connected the packing request schema with the frontend

+++ controllers/orderControllers.js
@@ -1,4 +1,5 @@
 const Order = require("../models/order");
+const User = require("../models/user");
 const jwt = require("jsonwebtoken");
 
 //isolate token from authorization header
@@ -11,41 +12,54 @@ const getTokenFrom = (request) => {
 };
 
 //get all orders
-const order_index = async (req, res) => {
-  try {
-    const orders = await Order.find({}).populate("user", { email: 1 });
-    res.status(200).json(orders);
-  } catch (error) {
-    res.status(404).json("No orders found");
-  }
-};
-
-//get order details
-const order_details = async (req, res) => {
-  const id = req.params.id;
-  const order = await Order.findById(id);
-
-  if (order) {
-    res.json(order);
-  } else {
-    res.status(404).end();
-  }
-};
-
-//create order
-const order_create = async (req, res) => {
-  const body = req.body;
-  //verify token
+const order_index = async (req, res, next) => {
   const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
   if (!decodedToken.id) {
     return res.status(401).json({ error: "Token invalid" });
   }
-  const user = await User.findById(body.userId);
+  try {
+    const orders = await Order.find({}).populate("user", { email: 1 });
+    if (orders) {
+      res.status(200).json(orders);
+    } else {
+      res.status(404).json("Orders not found");
+    }
+  } catch (error) {
+    next(error);
+  }
+};
+
+//get order details
+const order_details = async (req, res, next) => {
+  const id = req.params.id;
+
+  try {
+    const order = await Order.findById(id);
+    if (order) {
+      res.status(200).json(order);
+    } else {
+      res.status(404).json({ error: "Order not found" });
+    }
+  } catch (error) {
+    next(error);
+  }
+};
+
+//create order
+const order_create = async (req, res, next) => {
+  const body = req.body;
+
+  //verify token
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.JWT_SECRET);
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "Token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
   const order = new Order({
     user: user._id,
     customsInformation: body.customsInformation,
     selectedAddons: body.selectedAddons,
-    selectedCarrier: body.selectedCarrier,
+    //selectedCarrier: body.selectedCarrier,
   });
 
   try {
@@ -54,25 +68,37 @@ const order_create = async (req, res) => {
     await user.save();
     res.status(201).json(savedOrder);
   } catch (error) {
-    res.status(500).json(error);
+    next(error);
   }
 };
 
 //update an order
-const order_update = async (req, res) => {
+const order_update = async (req, res, next) => {
+  //verify token
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "Token invalid" });
+  }
   const id = req.params.id;
   const updateInfo = req.body;
   try {
-    const result = await Order.findByIdAndUpdate(id, updateInfo, { new: true });
-    await result.save();
-    res.status(201).json(result);
+    const updatedOrder = await Order.findByIdAndUpdate(id, updateInfo, {
+      new: true,
+    });
+
+    res.status(201).json(updatedOrder);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
 //delete an order
-const order_delete = async (req, res) => {
+const order_delete = async (req, res, next) => {
+  //verify token
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET);
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "Token invalid" });
+  }
   try {
     const order = await Order.findByIdAndDelete(req.params.id);
     if (!order) {
@@ -81,7 +107,7 @@ const order_delete = async (req, res) => {
       res.status(200).json("Successfully deleted parcel");
     }
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 module.exports = {
